backend/requirements.txt
# Backend dependencies
# WEEK 1
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
email-validator
sqlalchemy
psycopg2-binary
alembic
python-jose
python-multipart
passlib[argon2]
argon2-cffi
pytest
httpx

# WEEK 2
pymupdf
sentence-transformers
# transformers
# torch
faiss-cpu
numpy
scikit-learn

# # llm
# transformers
# torch
# accelerate
# sentencepiece

# experimental for changing the LLm inference
llama-cpp-python